requirements.txt
fastapi[standard]
uvicorn
pydantic
pydantic_settings
boto3
mangum
httpx
python-telegram-bot
pytest